+++ internal/service/database/database_db_systems_data_source.go
@@ -260,6 +260,8 @@ func (s *DatabaseDbSystemsDataSourceCrud) SetData() error {
 
 		dbSystem["security_attributes"] = tfresource.SecurityAttributesToMap(r.SecurityAttributes)
 
+		dbSystem["scan_ipv6ids"] = r.ScanIpv6Ids
+
 		if r.Shape != nil {
 			dbSystem["shape"] = *r.Shape
 		}
@@ -295,6 +297,8 @@ func (s *DatabaseDbSystemsDataSourceCrud) SetData() error {
 		}
 
 		dbSystem["vip_ids"] = r.VipIds
+
+		dbSystem["vipv6ids"] = r.Vipv6Ids
 
 		if r.ZoneId != nil {
 			dbSystem["zone_id"] = *r.ZoneId